+++ api_ecommerce/app/Http/Controllers/Admin/product/CategorieController.php
@@ -95,6 +95,13 @@ class CategorieController extends Controller
     public function destroy(string $id)
     {
         $categorie = Categorie::findOrFail($id);
+
+        if($categorie->product_categorie_firsts->count() > 0 ||
+                $categorie->product_categorie_secodns->count() > 0 ||
+                $categorie->product_categorie_thirds->count() > 0){
+          return response()->json(["message" => 403,"message_text" =>
+           "LA CATEGORIA YA ESTA RELACIONADO CON ALGUNOS O UN PRODUCTO"]);
+        }
         $categorie->delete();
         //validar que la categoria no este en ningun producto
         return response()->json(["message" => 200]);
